+++ .gitignore
@@ -1,0 +1,6 @@
+
+target/
+
+egov_project.iml
+
+.idea/
