+++ src/main/java/egovframework/example/sample/dao/ProgrammingMapper.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package egovframework.example.sample.service.impl;
+package egovframework.example.sample.dao;
 
 import java.util.List;
 
-import egovframework.example.sample.service.SampleDefaultVO;
-import egovframework.example.sample.service.SampleVO;
+import egovframework.example.sample.service.ProgrammingVO;
 
 import egovframework.rte.psl.dataaccess.mapper.Mapper;
 
@@ -38,7 +37,7 @@ import egovframework.rte.psl.dataaccess.mapper.Mapper;
  * </pre>
  */
 @Mapper("sampleMapper")
-public interface SampleMapper {
+public interface ProgrammingMapper {
 
 	/**
 	 * 글을 등록한다.
@@ -46,7 +45,7 @@ public interface SampleMapper {
 	 * @return 등록 결과
 	 * @exception Exception
 	 */
-	void insertSample(SampleVO vo) throws Exception;
+	void insertProgramming(ProgrammingVO vo) throws Exception;
 
 	/**
 	 * 글을 수정한다.
@@ -54,7 +53,7 @@ public interface SampleMapper {
 	 * @return void형
 	 * @exception Exception
 	 */
-	void updateSample(SampleVO vo) throws Exception;
+	void updateProgramming(ProgrammingVO vo) throws Exception;
 
 	/**
 	 * 글을 삭제한다.
@@ -62,7 +61,7 @@ public interface SampleMapper {
 	 * @return void형
 	 * @exception Exception
 	 */
-	void deleteSample(SampleVO vo) throws Exception;
+	void deleteProgramming(ProgrammingVO vo) throws Exception;
 
 	/**
 	 * 글을 조회한다.
@@ -70,7 +69,7 @@ public interface SampleMapper {
 	 * @return 조회한 글
 	 * @exception Exception
 	 */
-	SampleVO selectSample(SampleVO vo) throws Exception;
+	ProgrammingVO selectProgramming(ProgrammingVO vo) throws Exception;
 
 	/**
 	 * 글 목록을 조회한다.
@@ -78,7 +77,7 @@ public interface SampleMapper {
 	 * @return 글 목록
 	 * @exception Exception
 	 */
-	List<?> selectSampleList(SampleDefaultVO searchVO) throws Exception;
+	List<?> selectProgrammingList(ProgrammingVO vo) throws Exception;
 
 	/**
 	 * 글 총 갯수를 조회한다.
@@ -86,6 +85,5 @@ public interface SampleMapper {
 	 * @return 글 총 갯수
 	 * @exception
 	 */
-	int selectSampleListTotCnt(SampleDefaultVO searchVO);
 
 }
