+++ src/main/java/egovframework/example/sample/dao/SampleMapper.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package egovframework.example.sample.service.impl;
+package egovframework.example.sample.dao;
 
 import java.util.List;
 
-import egovframework.example.sample.service.ProgrammingVO;
+import egovframework.example.sample.service.SampleDefaultVO;
+import egovframework.example.sample.service.SampleVO;
 
 import egovframework.rte.psl.dataaccess.mapper.Mapper;
 
@@ -37,7 +38,7 @@ import egovframework.rte.psl.dataaccess.mapper.Mapper;
  * </pre>
  */
 @Mapper("sampleMapper")
-public interface ProgrammingMapper {
+public interface SampleMapper {
 
 	/**
 	 * 글을 등록한다.
@@ -45,7 +46,7 @@ public interface ProgrammingMapper {
 	 * @return 등록 결과
 	 * @exception Exception
 	 */
-	void insertProgramming(ProgrammingVO vo) throws Exception;
+	void insertSample(SampleVO vo) throws Exception;
 
 	/**
 	 * 글을 수정한다.
@@ -53,7 +54,7 @@ public interface ProgrammingMapper {
 	 * @return void형
 	 * @exception Exception
 	 */
-	void updateProgramming(ProgrammingVO vo) throws Exception;
+	void updateSample(SampleVO vo) throws Exception;
 
 	/**
 	 * 글을 삭제한다.
@@ -61,7 +62,7 @@ public interface ProgrammingMapper {
 	 * @return void형
 	 * @exception Exception
 	 */
-	void deleteProgramming(ProgrammingVO vo) throws Exception;
+	void deleteSample(SampleVO vo) throws Exception;
 
 	/**
 	 * 글을 조회한다.
@@ -69,7 +70,7 @@ public interface ProgrammingMapper {
 	 * @return 조회한 글
 	 * @exception Exception
 	 */
-	ProgrammingVO selectProgramming(ProgrammingVO vo) throws Exception;
+	SampleVO selectSample(SampleVO vo) throws Exception;
 
 	/**
 	 * 글 목록을 조회한다.
@@ -77,7 +78,7 @@ public interface ProgrammingMapper {
 	 * @return 글 목록
 	 * @exception Exception
 	 */
-	List<?> selectProgrammingList(ProgrammingVO vo) throws Exception;
+	List<?> selectSampleList(SampleDefaultVO searchVO) throws Exception;
 
 	/**
 	 * 글 총 갯수를 조회한다.
@@ -85,5 +86,6 @@ public interface ProgrammingMapper {
 	 * @return 글 총 갯수
 	 * @exception
 	 */
+	int selectSampleListTotCnt(SampleDefaultVO searchVO);
 
 }
