+++ src/main/resources/egovframework/spring/context-datasource.xml
@@ -3,37 +3,12 @@
 	xmlns:jdbc="http://www.springframework.org/schema/jdbc"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.0.xsd
         http://www.springframework.org/schema/jdbc  http://www.springframework.org/schema/jdbc/spring-jdbc-4.0.xsd">
-	
 
- 	<jdbc:embedded-database id="dataSource" type="HSQL"> 
- 		<jdbc:script location= "classpath:/db/sampledb.sql"/> 
-	</jdbc:embedded-database>
+	<bean name="dataSource" class="org.springframework.jdbc.datasource.DriverManagerDataSource">
+		<property name="driverClassName" value="org.mariadb.jdbc.Driver"/>
+		<property name="url" value="jdbc:mariadb://nas.pocketr.net:3307/taejung"/>
+		<property name="username" value="taejung"/>
+		<property name="password" value="bX4*Hjq!F7"/>
+	</bean>
 
-     
-     
-    <!-- hsql (테스트용 메모리 DB)
-    <bean id="dataSource" class="org.apache.commons.dbcp2.BasicDataSource" destroy-method="close">
-        <property name="driverClassName" value="net.sf.log4jdbc.DriverSpy"/>
-        <property name="url" value="jdbc:log4jdbc:hsqldb:hsql://localhost/sampledb"/>
-        <property name="username" value="sa"/>
-    </bean>
-    -->  
-    
-<!--     <bean id="dataSource" class="org.apache.commons.dbcp2.BasicDataSource" destroy-method="close"> -->
-<!--         <property name="driverClassName" value="com.mysql.jdbc.Driver"/> -->
-<!--         <property name="url" value="jdbc:mysql://127.0.0.1:3306/taejungdb" /> -->
-<!--         <property name="username" value="taejung"/> -->
-<!--         <property name="password" value="1234"/> -->
-<!--     </bean> -->
-
-    
-    <!-- oracle (POM에서 commons-dbcp, ojdbc(라이센스 사항으로 별도로 배포되지 않음) 관련 라이브러리 설정)
-    <bean id="dataSource" class="org.apache.commons.dbcp2.BasicDataSource" destroy-method="close">
-        <property name="driverClassName" value="oracle.jdbc.driver.OracleDriver"/>
-        <property name="url" value="jdbc:oracle:thin:@127.0.0.1:1521:example" />
-        <property name="username" value="user"/>
-        <property name="password" value="password"/>
-    </bean>
-    -->
-    
 </beans>
