+++ src/main/resources/egovframework/spring/context-mapper.xml
@@ -3,15 +3,15 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.0.xsd">
 
 	<!-- SqlSession setup for MyBatis Database Layer -->
-	<bean id="sqlSession" class="org.mybatis.spring.SqlSessionFactoryBean">
+	<bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
 		<property name="dataSource" ref="dataSource" />
 		<property name="configLocation" value="classpath:/egovframework/sqlmap/example/sql-mapper-config.xml" />
-		<property name="mapperLocations" value="classpath:/egovframework/sqlmap/example/mappers/*.xml" />
+		<property name="mapperLocations" value="classpath*:/egovframework/sqlmap/example/mappers/*.xml" />
 	</bean>
 
 	<!-- MapperConfigurer setup for MyBatis Database Layer with @Mapper("deptMapper") in DeptMapper Interface -->
- 	<bean class="egovframework.rte.psl.dataaccess.mapper.MapperConfigurer">
-		<property name="basePackage" value="egovframework.example.sample.service.impl" />
-	</bean>
+<!-- 	<bean class="egovframework.rte.psl.dataaccess.mapper.MapperConfigurer">-->
+<!--		<property name="basePackage" value="egovframework.example.sample.service.impl" />-->
+<!--	</bean>-->
     
 </beans>
