+++ src/main/resources/egovframework/sqlmap/example/mappers/EgovSample_Sample_SQL.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="egovframework.example.sample.service.impl.SampleMapper">
+<mapper namespace="egovframework.example.sample.dao.SampleMapper">
 
 	<resultMap id="sample" type="egovframework.example.sample.service.SampleVO">
 		<result property="id" column="id"/>
@@ -53,41 +53,41 @@
 
 	</select>
 
-	<select id="selectSampleList" parameterType="searchVO" resultType="egovMap">
+<!--	<select id="selectSampleList" parameterType="searchVO" resultType="egovMap">-->
 
-			SELECT
-				ID, NAME, DESCRIPTION, USE_YN, REG_USER
-			FROM SAMPLE
-			WHERE 1=1
-			<if test="searchKeyword != null and searchKeyword != ''">
-		        <choose>
-		            <when test="searchCondition == 0">
-						AND	ID LIKE '%' || #{searchKeyword} || '%'
-					</when>
-		            <when test="searchCondition == 1">
-						AND	NAME LIKE '%' || #{searchKeyword} || '%'
-					</when>
-				</choose>
-			</if>
-			ORDER BY ID ASC
-			LIMIT #{recordCountPerPage} OFFSET #{firstIndex}
-	</select>
+<!--			SELECT-->
+<!--				ID, NAME, DESCRIPTION, USE_YN, REG_USER-->
+<!--			FROM SAMPLE-->
+<!--			WHERE 1=1-->
+<!--			<if test="searchKeyword != null and searchKeyword != ''">-->
+<!--		        <choose>-->
+<!--		            <when test="searchCondition == 0">-->
+<!--						AND	ID LIKE '%' || #{searchKeyword} || '%'-->
+<!--					</when>-->
+<!--		            <when test="searchCondition == 1">-->
+<!--						AND	NAME LIKE '%' || #{searchKeyword} || '%'-->
+<!--					</when>-->
+<!--				</choose>-->
+<!--			</if>-->
+<!--			ORDER BY ID ASC-->
+<!--			LIMIT #{recordCountPerPage} OFFSET #{firstIndex}-->
+<!--	</select>-->
 
-	<select id="selectSampleListTotCnt" parameterType="searchVO" resultType="int">
+<!--	<select id="selectSampleListTotCnt" parameterType="searchVO" resultType="int">-->
 
-			SELECT COUNT(*) totcnt
-			FROM SAMPLE
-			WHERE 1=1
-			<if test="searchKeyword != null and searchKeyword != ''">
-		        <choose>
-		            <when test="searchCondition == 0">
-						AND ID LIKE '%' || #{searchKeyword} || '%'
-					</when>
-		            <when test="searchCondition == 1">
-						AND	NAME LIKE '%' || #{searchKeyword} || '%'
-					</when>
-				</choose>
-			</if>
-	</select>
+<!--			SELECT COUNT(*) totcnt-->
+<!--			FROM SAMPLE-->
+<!--			WHERE 1=1-->
+<!--			<if test="searchKeyword != null and searchKeyword != ''">-->
+<!--		        <choose>-->
+<!--		            <when test="searchCondition == 0">-->
+<!--						AND ID LIKE '%' || #{searchKeyword} || '%'-->
+<!--					</when>-->
+<!--		            <when test="searchCondition == 1">-->
+<!--						AND	NAME LIKE '%' || #{searchKeyword} || '%'-->
+<!--					</when>-->
+<!--				</choose>-->
+<!--			</if>-->
+<!--	</select>-->
 
 </mapper>
