+++ src/main/resources/egovframework/sqlmap/example/sample/EgovSample_Sample_SQL.xml
@@ -57,33 +57,33 @@
         ]]>
     </select>
 	
-	<select id="sampleDAO.selectSampleList" parameterClass="searchVO" resultClass="egovMap">
+<!--	<select id="sampleDAO.selectSampleList" parameterClass="searchVO" resultClass="egovMap">-->
 
-			SELECT
-				ID, NAME, DESCRIPTION, USE_YN, REG_USER
-			FROM SAMPLE
-			WHERE 1=1
-			<isEqual prepend="AND" property="searchCondition" compareValue="0">
-				ID LIKE '%' || #searchKeyword# || '%'
-			</isEqual>
-			<isEqual prepend="AND" property="searchCondition" compareValue="1">
-				NAME LIKE '%' || #searchKeyword# || '%'
-			</isEqual>
-			ORDER BY ID DESC
-			LIMIT #recordCountPerPage# OFFSET #firstIndex#		
-	</select>
-	
-	<select id="sampleDAO.selectSampleListTotCnt" parameterClass="searchVO" resultClass="int">
+<!--			SELECT-->
+<!--				ID, NAME, DESCRIPTION, USE_YN, REG_USER-->
+<!--			FROM SAMPLE-->
+<!--			WHERE 1=1-->
+<!--			<isEqual prepend="AND" property="searchCondition" compareValue="0">-->
+<!--				ID LIKE '%' || #searchKeyword# || '%'-->
+<!--			</isEqual>-->
+<!--			<isEqual prepend="AND" property="searchCondition" compareValue="1">-->
+<!--				NAME LIKE '%' || #searchKeyword# || '%'-->
+<!--			</isEqual>-->
+<!--			ORDER BY ID DESC-->
+<!--			LIMIT #recordCountPerPage# OFFSET #firstIndex#		-->
+<!--	</select>-->
+<!--	-->
+<!--	<select id="sampleDAO.selectSampleListTotCnt" parameterClass="searchVO" resultClass="int">-->
 
-			SELECT COUNT(*) totcnt
-			FROM SAMPLE
-			WHERE 1=1
-			<isEqual prepend="AND" property="searchCondition" compareValue="0">
-				ID LIKE '%' || #searchKeyword# || '%'
-			</isEqual>
-			<isEqual prepend="AND" property="searchCondition" compareValue="1">
-				NAME LIKE '%' || #searchKeyword# || '%'
-			</isEqual>
-	</select>
+<!--			SELECT COUNT(*) totcnt-->
+<!--			FROM SAMPLE-->
+<!--			WHERE 1=1-->
+<!--			<isEqual prepend="AND" property="searchCondition" compareValue="0">-->
+<!--				ID LIKE '%' || #searchKeyword# || '%'-->
+<!--			</isEqual>-->
+<!--			<isEqual prepend="AND" property="searchCondition" compareValue="1">-->
+<!--				NAME LIKE '%' || #searchKeyword# || '%'-->
+<!--			</isEqual>-->
+<!--	</select>-->
 
 </sqlMap>
