+++ src/main/resources/egovframework/sqlmap/example/sql-mapper-config.xml
@@ -3,10 +3,9 @@
 
 <configuration>
     <typeAliases>
-<!-- 		<typeAlias alias="egovMap" type="egovframework.rte.psl.dataaccess.util.EgovMap"/> -->
-<!-- 		<typeAlias alias="searchVO" type="egovframework.example.sample.service.SampleDefaultVO"/> -->
-<!-- 		<typeAlias alias="sampleVO" type="egovframework.example.sample.service.SampleVO"/> -->
 		<typeAlias alias="egovMap" type="egovframework.rte.psl.dataaccess.util.EgovMap"/>
 		<typeAlias alias="ProgrammingVO" type="egovframework.example.sample.service.ProgrammingVO"/>
+		<typeAlias alias="SampleDefaultVO" type="egovframework.example.sample.service.SampleDefaultVO"/>
+		<typeAlias alias="SampleVO" type="egovframework.example.sample.service.SampleVO"/>
     </typeAliases>
 </configuration>
